somewhat stable formant filter (at A and E vowels)

+++ src/jni/formant_filter.h
@@ -1,7 +1,8 @@
 /**
  * The MIT License (MIT)
  *
- * Copyright (c) 2013-2014 Igor Zinken - http://www.igorski.nl
+ * based on public source code by alex@smartelectronix.com, adapted
+ * by Igor Zinken - http://www.igorski.nl
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
  * this software and associated documentation files (the "Software"), to deal in
@@ -30,8 +31,10 @@ class FormantFilter : public BaseProcessor
     public:
         FormantFilter( float aVowel );
         ~FormantFilter();
+
         void setVowel( float aVowel );
         float getVowel();
+
         void process( AudioBuffer* audioBuffer, bool isMonoSource );
         bool isCacheable();
 
@@ -48,7 +51,7 @@ class FormantFilter : public BaseProcessor
         static const int VOWEL_U = 4;
 
     private:
-        double  _vowel;
+        float  _vowel;
         double _currentCoeffs[ 11 ];
         double _coeffs[ 5 ][ 11 ];
         double _memory[ 10 ];
